Move ReversedColor to main entry point
