+++ packages/components/src/LinkButton/LinkButton.tsx
@@ -2,7 +2,7 @@ import React, { forwardRef } from 'react'
 import { Link as RACLink, type LinkProps as RACLinkProps } from 'react-aria-components'
 import { type ButtonUIProps } from '~components/__next__/Button'
 import { ButtonContent } from '~components/__next__/Button/subcomponents'
-import { useReversedColors } from '~components/__utilities__/v3'
+import { useReversedColors } from '~components/utils'
 import { mergeClassNames } from '~components/utils/mergeClassNames'
 import buttonStyles from '../__next__/Button/Button.module.css'
 import styles from './LinkButton.module.css'
