+++ packages/components/src/LinkButton/_docs/LinkButton--api-specification.mdx
@@ -82,7 +82,7 @@ To enable the reversed theme, you will need to wrap the component or application
 ```tsx
 import { RouterProvider } from 'react-aria-components'
 import { Button } from '@kaizen/components/next'
-import { ReversedColors } from '@kaizen/components/v3/utilities'
+import { ReversedColors } from '@kaizen/components'
 // application code
 
 return (
