+++ packages/components/src/__next__/Button/_docs/Button--api-specification.mdx
@@ -70,7 +70,7 @@ To enable the reversed theme, you will need to wrap the component or application
 
 ```tsx
 import { Button } from '@kaizen/components/next'
-import { ReversedColors } from '@kaizen/components/v3/utilities'
+import { ReversedColors } from '@kaizen/components'
 // application code
 
 return (
