+++ packages/components/src/__next__/Select/subcomponents/ListBox/ListBox.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useRef, type HTMLAttributes, type Key, type ReactNode
 import { useListBox, type AriaListBoxOptions } from '@react-aria/listbox'
 import { type SelectState } from '@react-stately/select'
 import classnames from 'classnames'
-import { useIsClientReady } from '~components/__utilities__/useIsClientReady'
 import { type OverrideClassName } from '~components/types/OverrideClassName'
+import { useIsClientReady } from '~components/utils/useIsClientReady'
 import { useSelectContext } from '../../context'
 import { type SelectItem, type SelectOption } from '../../types'
 import styles from './ListBox.module.scss'
