+++ packages/components/src/__next__/Tooltip/OverlayArrow.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { OverlayArrow as RACOverlayArrow, type OverlayArrowProps } from 'react-aria-components'
-import { useReversedColors } from '~components/__utilities__/v3'
+import { useReversedColors } from '~components/utils'
 import { mergeClassNames } from '~components/utils/mergeClassNames'
 
 import styles from './OverlayArrow.module.scss'
