+++ packages/components/src/__utilities__/index.ts
@@ -1,0 +1,2 @@
+// Remove this folder in 2.0 release when v3/utilities entry point is removed
+export * from '../utils/ReversedColors'
