--- packages/components/src/__utilities__/v3.ts
@@ -1,1 +1,0 @@
-export * from './ReversedColors/v3'
