+++ packages/components/src/utils/index.ts
@@ -1,3 +1,4 @@
 // Not all utils are exported because most are intended for internal use only
 export * from './useMediaQueries'
 export * from './hostedAssets'
+export * from './ReversedColors'
