+++ packages/components/v3/utilities/package.json
@@ -1,5 +1,5 @@
 {
   "main": "../../dist/cjs/utilitiesV3.cjs",
   "module": "../../dist/esm/utilitiesV3.mjs",
-  "types": "../../dist/types/__utilities__/v3.d.ts"
+  "types": "../../dist/types/__utilities__/index.d.ts"
 }
